fix: enabled에 따른 예외 처리

+++ src/main/java/com/iscourse/api/service/CourseService.java
@@ -2,12 +2,14 @@ package com.iscourse.api.service;
 
 import com.iscourse.api.domain.course.Course;
 import com.iscourse.api.domain.course.CoursePlace;
+import com.iscourse.api.domain.course.Place;
 import com.iscourse.api.domain.course.dto.CourseShareDto;
 import com.iscourse.api.domain.member.Member;
 import com.iscourse.api.domain.member.MemberCourse;
 import com.iscourse.api.domain.member.MemberCourseLike;
 import com.iscourse.api.domain.member.MemberRoleType;
 import com.iscourse.api.exception.DuplicateCourseException;
+import com.iscourse.api.exception.UnavailableEntityException;
 import com.iscourse.api.repository.course.CoursePlaceRepository;
 import com.iscourse.api.repository.course.CourseRepository;
 import com.iscourse.api.repository.course.PlaceRepository;
@@ -36,8 +38,12 @@ public class CourseService {
         // 좋아요 기능 구현
         Member member = memberRepository.findById(memberId).orElseThrow(IllegalArgumentException::new);
         Course course = courseRepository.findById(courseId).orElseThrow(IllegalArgumentException::new);
+
+        if (!course.getEnabled()) {
+            throw new UnavailableEntityException("비활성화된 코스입니다.");
+        }
+
         Optional<MemberCourseLike> memberCourseLike = memberCourseLikeRepository.findByMemberAndCourse(member, course);
-        System.out.println("memberCourseLike = " + memberCourseLike);
         if (!memberCourseLike.isEmpty()) {
             memberCourseLikeRepository.delete(memberCourseLike.get());
             course.subtractLike();
@@ -51,6 +57,11 @@ public class CourseService {
     public void selectCourse(Long courseId, Long memberId) {
         Member member = memberRepository.findById(memberId).orElseThrow(IllegalArgumentException::new);
         Course course = courseRepository.findById(courseId).orElseThrow(IllegalArgumentException::new);
+
+        if (!course.getEnabled()) {
+            throw new UnavailableEntityException("비활성화된 코스입니다.");
+        }
+
         memberCourseRepository.findByMemberAndCourse(member, course).ifPresent(memberCourse -> {
             throw new DuplicateCourseException("이미 등록된 코스입니다.");
         });
@@ -69,9 +80,15 @@ public class CourseService {
         );
         courseRepository.save(course);
         for (int i = 0; i < courseShareDto.getPlaceIdList().size(); i++) {
+            Place place = placeRepository.findById(courseShareDto.getPlaceIdList().get(i)).orElseThrow(IllegalArgumentException::new);
+
+            if (!place.getEnabled()) {
+                throw new UnavailableEntityException("비활성화된 장소입니다.");
+            }
+
             CoursePlace coursePlace = new CoursePlace(
                     course,
-                    placeRepository.findById(courseShareDto.getPlaceIdList().get(i)).orElseThrow(IllegalArgumentException::new),
+                    place,
                     i
             );
             coursePlaceRepository.save(coursePlace);
